requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
black==24.4.0
click==8.1.7
colorama==0.4.6
et-xmlfile==1.1.0
fastapi==0.110.1
flake8==7.0.0
Flake8-pyproject==1.2.3
greenlet==3.0.3
h11==0.14.0
idna==3.7
isort==5.13.2
Mako==1.3.3
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydantic==2.7.0
pydantic_core==2.18.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
tzdata==2024.1
uvicorn==0.29.0
